Fixed according to mzhang's pull request review (https://github.com/DxChainNetwork/godx/pull/58)
1. added comments: common/writeaheadlog/recover.go:20
2. Moved data structure outside function: common/writeaheadlog/recover.go:12
3. The txnStatusInvalid has value 0, and when creating a new structure, the status is default to this value. That's why it is unused before. Now it is used: common/writeaheadlog/transaction.go:71
4. txnStatusWritten was a reasonable name. But from the perspective make the naming more clear to all users, I agreed with modifying the naming field and renamed the status to txnStatusUncommitted

Question:

Although not used, transaction.ID should be accessed publicly. This is because users might make a sequential transactions and user could only proceed when multiple transactions available. Thus the transaction ID is needed, which is incremented everytime user creates a new transaction. Your question reminds me to modify the transaction ID. Previously, the transaction ID is assigned during commit. Now, it is assigned when NewTransaction.